Refactored the CAMUS results data structures to be more generic
Also made public some previously private functions to load+filter `Group` objects from `h5py.Group`s
